Check operator when creating UNIQUE index on PARTITION table
Currently, we only check if every column in the partition key
is included in the unique index. But this is not enough when
custom opclass exists. Since there can be cases, that some
values belong to different partitions should be considered
equal according to unique index's opclass.

So, we should check the compatiblity of equality operator
between partition key's opclass and index column's opclass.
Only if they are the same operator should we consider them as
compatible.